+++ src/base/common/pod.h
@@ -16,6 +16,6 @@
 #include <type_traits>
 
 #define WH_IS_POD(X) (std::is_pod<X>::value)
-#define WH_POD_ASSERT(X) WH_STATIC_ASSERT(std::is_pod<X>::value, #X" should be POD (Plain Old Data) type")
+#define WH_POD_ASSERT(X) WH_STATIC_ASSERT(WH_IS_POD(X), #X" should be POD (Plain Old Data) type")
 
 #endif /* WH_BASE_COMMON_POD_H_ */