+++ src/hub/Socket.h
@@ -93,9 +93,9 @@ public:
 	 */
 	int shutdown(int how = SHUT_RDWR) noexcept;
 	/*
-	 * Returns the number of bytes read, possibly zero (buffer full or would block),
-	 * or -1 if the connection has been closed cleanly. Clears the read IO event from
-	 * this connection if the call would block.
+	 * Returns the number of bytes read, possibly zero (buffer full or would
+	 * block), or -1 if the connection has been closed cleanly. Clears out the
+	 * read IO event from this connection if the call would block.
 	 */
 	ssize_t read();
 	/*