+++ docs/ChangeLog.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- **Hub::retainMessage** sets the Message flag to **MSG_WAIT_PROCESSING**.
+- **Inotifier::read** doesn't unnecessarily zero out the IO buffer.
+
 ## [3.0.0] - 2021-03-02
 
 ### Added